fixed: getVisibleOffset() type should be int

+++ library/src/main/java/com/mypopsy/widget/SlidingUpPaneLayout.java
@@ -213,7 +213,7 @@ public class SlidingUpPaneLayout extends ViewGroup {
         return mSlideOffset;
     }
 
-    public float getVisibleOffset() {
+    public int getVisibleOffset() {
         return mVisibleOffset;
     }
